Kept the shared session stack on both fake data sources: home no longer
serves sessions, and reports derives them from FakeInterviewSessions.

+++ feature/practicesession/src/main/java/com/iti/careerpilot/practicesession/domain/models/SessionResult.kt
@@ -23,7 +23,7 @@ data class SessionQuestionResult(
     val userTranscript: String,
     val durationMs: Long,
     val speechRateWpm: Double,
-    val avgPauseMs: Int,
+    val avgPauseMs: Double,
     val silenceRatio: Double,
     val createdAt: String,
     val completedAt: String,
